chore: moved menuToggleButton into MapControls to better control css
